Friend view wired, user model updates, hack to allow you to friend people
I needed a way to actually friend people in order to test some of this,
so I added a property you can set to make the FriendVC into an "Everyone" VC
and then clicking friends them.  Not pretty, but functional.

Rest of this is just the wiring and model stuff for friends.